+++ library/android-mvc-test/src/main/java/com/shipdream/lib/android/mvc/view/injection/InjectionTestActivityTestRootFragment.java
@@ -18,7 +18,6 @@ package com.shipdream.lib.android.mvc.view.injection;
 
 import android.os.Bundle;
 
-import com.shipdream.lib.android.mvc.Controller;
 import com.shipdream.lib.android.mvc.FragmentController;
 import com.shipdream.lib.android.mvc.MvcFragment;
 import com.shipdream.lib.android.mvc.TestActivity;
@@ -29,14 +28,15 @@ import com.shipdream.lib.android.mvc.view.injection.controller.ControllerD;
 
 public class InjectionTestActivityTestRootFragment extends TestActivity {
     @Override
-    protected Class<? extends MvcFragment> mapControllerFragment(Class<? extends Controller> presenterClass) {
-        if (presenterClass == ControllerA.class) {
+    protected Class<? extends MvcFragment> mapFragmentRouting(
+            Class<? extends FragmentController> controllerClass) {
+        if (controllerClass == ControllerA.class) {
             return FragmentA.class;
-        } else if (presenterClass == ControllerB.class) {
+        } else if (controllerClass == ControllerB.class) {
             return FragmentB.class;
-        } else if (presenterClass == ControllerC.class) {
+        } else if (controllerClass == ControllerC.class) {
             return FragmentC.class;
-        } else if (presenterClass == ControllerD.class) {
+        } else if (controllerClass == ControllerD.class) {
             return FragmentD.class;
         }
         return null;